Refactor app structure into separate components
