+++ server/index.js
@@ -1,3 +1,4 @@
+/* eslint-env node */
 import express from 'express'
 import cors from 'cors'
 import Database from 'better-sqlite3'
